move function to _helpers
