+++ scripts/cluster_network.py
@@ -134,9 +134,9 @@ from _helpers import (
     REGION_COLS,
     configure_logging,
     create_logger,
-    get_base_carrier,
     locate_bus,
     nearest_shape,
+    restore_base_carrier_names,
     update_config_dictionary,
     update_p_nom_max,
 )
@@ -618,38 +618,6 @@ def cluster_regions(busmaps, inputs, output):
         regions_c.index.name = "name"
         regions_c = regions_c.reset_index()
         regions_c.to_file(getattr(output, which))
-
-
-def restore_base_carrier_names(n: pypsa.Network) -> None:
-    """
-    Restore carrier names from carrier_gy format (e.g., "solar-2020") to base carrier (e.g., "solar").
-
-    This is called after all aggregation operations to clean up carrier names while
-    preserving build year information in component names/indices.
-
-    Generator indices keep build year information (e.g., "US0 1 solar-2025"), but carrier becomes base ("solar").
-
-    Parameters
-    ----------
-    n : pypsa.Network
-        The PyPSA network to modify in-place.
-    """
-    # Restore base carrier names for generators
-    n.generators["carrier"] = n.generators["carrier"].apply(get_base_carrier)
-
-    # Restore base carrier names for storage units
-    n.storage_units["carrier"] = n.storage_units["carrier"].apply(get_base_carrier)
-
-    # Remove year-tagged carriers
-    year_tagged_carriers = []
-    for carrier in n.carriers.index:
-        parts = carrier.rsplit("-", 1)
-        if len(parts) == 2 and parts[1].isdigit() and len(parts[1]) == 4:
-            year_tagged_carriers.append(carrier)
-
-    if len(year_tagged_carriers) > 0:
-        n.mremove("Carrier", year_tagged_carriers)
-        logger.info(f"Removed year-tagged carriers: {', '.join(year_tagged_carriers)}")
 
 
 def replace_components(n, c, df, pnl):
